Refactor: Improve JSON storage and data types
late night debugging and now everything works

This commit refactors the JSON storage mechanism and related data types for improved clarity and functionality.

**Key changes:**

*   **Renamed `AutoDataStorageUtils` to `FtcJsonStorage`:** This provides a more descriptive name for the class responsible for JSON file operations.
*   **Introduced `FtcJsonExample.java`:** A new example OpMode demonstrating how to use `FtcJsonStorage` to read and write data.
*   **Enhanced `FtcJsonStorage`:**
    *   Added `getInternalFileNames()` to list files in internal storage.
    *   Made `getJsonFromMotorPositions` and `getJsonFromArguments` private as they are internal helper methods.
    *   Corrected parsing logic in `readAndParseAutoData` to properly handle nested JSON arrays for motor positions.
*   **Improved `DecodeDataTypes`:**
    *   Added default constructors to `ArtifactSequence`, `Coords`, and `DateMs`.
    *   Added a static factory method `ArtifactSequence.of()` for convenient instantiation.
    *   Added a constructor to `Coords` for direct initialization.
    *   Enhanced error handling in `ArtifactSequence.setSequence()` and `toStringArray()` for invalid colors.
    *   Simplified `MotorPositions.getEntries()` by removing the unnecessary index parameter.
*   **Updated `AprilTags.java`:**
    *   Integrated `FtcJsonStorage` for reading and writing autonomous data.
    *   Replaced the local `ArtifactColors` enum with `ArtifactColor` from `DecodeDataTypes`.
    *   Updated telemetry to display data read from JSON.
    *   Removed `@SuppressLint("DefaultLocale")`.
