+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/utils/DecodeDataTypes.java
@@ -16,12 +16,13 @@ public class DecodeDataTypes {
     public static class ArtifactSequence {
         private ArtifactColor[] sequence = new ArtifactColor[3];
 
-        public ArtifactSequence() {
-
-        }
-
+        public ArtifactSequence() {}
         public ArtifactSequence(ArtifactColor[] sequence) {
             this.sequence = sequence;
+        }
+
+        public static ArtifactSequence of(ArtifactColor... colors) {
+            return new ArtifactSequence(colors);
         }
 
         public ArtifactColor[] getSequence() {
@@ -32,12 +33,17 @@ public class DecodeDataTypes {
             if (sequence.length != 3) {
                 throw new IllegalArgumentException("Sequence must have 3 elements");
             }
+
             for (int i = 0; i < 3; i++) {
                 switch (sequence[i]) {
                     case "GREEN":
                         this.sequence[i] = ArtifactColor.GREEN;
+                        break;
                     case "PURPLE":
                         this.sequence[i] = ArtifactColor.PURPLE;
+                        break;
+                    default:
+                        throw new IllegalArgumentException("Invalid artifact color: " + sequence[i]);
                 }
             }
         }
@@ -49,8 +55,12 @@ public class DecodeDataTypes {
                 switch (sequence[i]) {
                     case GREEN:
                         stringArray[i] = "GREEN";
+                        break;
                     case PURPLE:
                         stringArray[i] = "PURPLE";
+                        break;
+                    default:
+                        throw new IllegalArgumentException("Invalid artifact color at index " + i + ": " + sequence[i]);
                 }
             }
             return stringArray;
@@ -64,6 +74,17 @@ public class DecodeDataTypes {
         private double pitch;
         private double roll;
         private double yaw;
+
+        public Coords() {}
+
+        public Coords(double x, double y, double z, double pitch, double roll, double yaw) {
+            this.x = x;
+            this.y = y;
+            this.z = z;
+            this.pitch = pitch;
+            this.roll = roll;
+            this.yaw = yaw;
+        }
 
         public void setCoords(double x, double y, double z, double pitch, double roll, double yaw) {
             this.x = x;
@@ -101,6 +122,8 @@ public class DecodeDataTypes {
 
     public static class DateMs {
         private long date_ms;
+
+        public DateMs() {}
 
         public DateMs(long init_ms) {
             date_ms = init_ms;
@@ -165,12 +188,8 @@ public class DecodeDataTypes {
             }
         }
 
-        public List<Map.Entry<String, Double>> getEntries(int n) {
+        public List<Map.Entry<String, Double>> getEntries() {
             if (motorPositions instanceof LinkedHashMap) {
-                if (n < 0 || n >= motorPositions.size()) {
-                    throw new IndexOutOfBoundsException("Index " + n + " is out of bounds for map size " + motorPositions.size());
-                }
-
                 return new ArrayList<>(motorPositions.entrySet());
             } else {
                 throw new IllegalStateException("MotorPositions needs to be a LinkedHashMap");
